Mango Types Classification System/requirements.txt
numpy==1.26.4
joblib==1.4.0
streamlit==1.33.0
Pillow==10.3.0
scikit-image==0.23.2
